Rename poolswap interval to pool.swap
