+++ src/module.indexer/model/dftx/pool.swap.interval.ts
